requirements.txt
fastapi
uvicorn[standard]
gTTS
redis
python-dotenv
requests
pydub
cloudinary